moves space to Revit wrappers
