+++ BRPLUSA.Domain.Core/BRPLUSA.Domain.Core.csproj
@@ -98,7 +98,7 @@
     <Compile Include="Interfaces\IEntity.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="WorksharingMonitorService.cs" />
-    <Compile Include="Wrappers\Space.cs" />
+    <Compile Include="Entities\Wrappers\Revit\Space.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
